Update Flu Image + Text Edits
Updated the influenza figure .png file to be the most recent one, and amended the rmd code to call that figure, as well as mention influenza B virus at the request of the Delatolla lab. Also removed text from a previous edit which explained the method modification on June 8th.