Remove VOC Figure
Remove VOC figure, as it is very outdated. As we look to transition responsibility of 613covid to another group, additional work to replace this figure is not reasonable.

Left some commented text to hold a spot on the page, however, as Tyson indicated he might prepare some text on general variant presence and levels. No figure will be coming back, however.