Change to "date" process, debugging statements
Commented out the previous approach to generating the start date for the modelling as it was throwing a error in the GH Actions environment. Shifted to the lubridate approach to generate the same date without that issue (hopefully)

Also added debugging statements to the end to test if things worked well.